Change the return type of call counts to long, int max value is not large enough

+++ bgee-core/src/main/java/org/bgee/model/expressiondata/call/ExpressionCallLoader.java
@@ -222,7 +222,7 @@ public class ExpressionCallLoader extends CommonService {
         
     }
 
-    public int loadDataCount() {
+    public long loadDataCount() {
         log.traceEntry();
         return log.traceExit(this.globalExprCallDAO.getGlobalExpressionCallsCount(
                 this.processedFilter.getDaoFilters()));
